move the LinearSigmoid test model to its own file
